[FEAT] - Property Service: Add Redis for API CRUD PropertyInteraction and create, get Property

+++ property-service/src/repositories/propertyInteraction.repository.ts
@@ -1,5 +1,10 @@
 import { UserInteractionType } from '@prisma/client';
-import { IPropertyInteractionReq, IPropertyInteractionRes } from '../interfaces/propertyInteraction';
+import {
+    IPropertyInteractionDeleteReq,
+    IPropertyInteractionReq,
+    IPropertyInteractionRes,
+    IPropertyInteractionUpdateReq,
+} from '../interfaces/propertyInteraction';
 import prisma from '../prisma/prismaClient';
 import CustomError from '../utils/error.util';
 
@@ -23,29 +28,33 @@ export const getPropertyInteractionById = async (interaction_id: string): Promis
 };
 
 export const updatePropertyInteraction = async ({
+    user_id,
     interaction_type,
     interaction_id,
-}: {
-    interaction_type: UserInteractionType;
-    interaction_id: string;
-}): Promise<IPropertyInteractionRes> => {
+}: IPropertyInteractionUpdateReq): Promise<IPropertyInteractionRes> => {
     return prisma.userPropertyInteractions.update({
-        where: { interaction_id },
+        where: { interaction_id, user_id },
         data: {
             interaction_type,
         },
     });
 };
 
-export const softDeletePropertyInteraction = async (interaction_id: string): Promise<IPropertyInteractionRes> => {
+export const softDeletePropertyInteraction = async ({
+    interaction_id,
+    user_id,
+}: IPropertyInteractionDeleteReq): Promise<IPropertyInteractionRes> => {
     return prisma.userPropertyInteractions.update({
-        where: { interaction_id },
+        where: { interaction_id, user_id },
         data: {
             deleted: true,
         },
     });
 };
 
-export const deletePropertyInteraction = async (interaction_id: string): Promise<IPropertyInteractionRes> => {
-    return prisma.userPropertyInteractions.delete({ where: { interaction_id } });
+export const deletePropertyInteraction = async ({
+    interaction_id,
+    user_id,
+}: IPropertyInteractionDeleteReq): Promise<IPropertyInteractionRes> => {
+    return prisma.userPropertyInteractions.delete({ where: { interaction_id, user_id } });
 };
